requirements.txt
Django==1.8.4
django-piston3==0.3rc2
